+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.0.0 - 2025-07-12
 
+- feat!(Fetch): removed `Fetch` component
 - feat!(Store): removed `has` function for better performance
 - feat!(useFetch): removed `reloading` return, you should use `loading` return instead
 - feat(CacheStore): new store implemented with [Cache API](https://developer.mozilla.org/docs/Web/API/Cache)