+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - feat!(Store): removed `has` function for better performance
 - feat!(useFetch): removed `reloading` return, you should use `loading` return instead
 - feat(CacheStore): new store implemented with [Cache API](https://developer.mozilla.org/docs/Web/API/Cache)
+- fix(useFetch): initial loading to be true
 
 ## 1.15.0 - 2025-07-12