+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.0.1 - 2025-12-15
+
+- fix(react): support React 16 without jsx-runtime
+
 ## 2.0.0 - 2025-07-12
 
 - feat!(DataStatus): removed `DataStatus` enum